requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via django
backports.entry-points-selectable==1.1.0
    # via virtualenv
black==21.9b0
    # via -r requirements.in
cfgv==3.3.1
    # via pre-commit
click==8.0.3
    # via black
distlib==0.3.3
    # via virtualenv
django==3.2.8
    # via -r requirements.in
filelock==3.3.2
    # via virtualenv
flake8==4.0.1
    # via -r requirements.in
identify==2.3.1
    # via pre-commit
isort==5.9.3
    # via -r requirements.in
markdown==3.3.4
    # via -r requirements.in
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
pathspec==0.9.0
    # via black
platformdirs==2.4.0
    # via
    #   black
    #   virtualenv
pre-commit==2.15.0
    # via -r requirements.in
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pytz==2021.3
    # via django
pyyaml==6.0
    # via
    #   -r requirements.in
    #   pre-commit
regex==2021.10.23
    # via black
six==1.16.0
    # via virtualenv
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via pre-commit
tomli==1.2.2
    # via
    #   -r requirements.in
    #   black
typing-extensions==3.10.0.2
    # via black
virtualenv==20.9.0
    # via pre-commit
